+++ app/models/user.rb
@@ -1,6 +1,9 @@
 # frozen_string_literal: true
 
 class User < ApplicationRecord
+  include Recoverable
+  include Rememberable
+
   enum role: { basic: 0, moderator: 1, admin: 2 }, _suffix: :role
   attr_accessor :old_password, :remember_token, :admin_edit
 
@@ -27,26 +30,6 @@ class User < ApplicationRecord
 
   def author?(obj)
     obj.user == self
-  end
-
-  def remember_me
-    self.remember_token = SecureRandom.urlsafe_base64
-    # rubocop:disable Rails/SkipsModelValidations
-    update_column :remember_token_digest, digest(remember_token)
-    # rubocop:enable Rails/SkipsModelValidations
-  end
-
-  def forget_me
-    # rubocop:disable Rails/SkipsModelValidations
-    update_column :remember_token_digest, nil
-    # rubocop:enable Rails/SkipsModelValidations
-    self.remember_token = nil
-  end
-
-  def remember_token_authenticated?(remember_token)
-    return false if remember_token_digest.blank?
-
-    BCrypt::Password.new(remember_token_digest).is_password?(remember_token)
   end
 
   private
